test_requirements.txt
coverage
django-coverage
ipdb
flake8
fabric3
mixer
mock
tox
